Attempted fix for ws hate resets: 1. Clamp CE and VE in UpdateEnmityFromDamage. 2. Make UpdateEnmityFromDamage use the same typing for damage argument that battleutils did when calling it.

+++ src/map/enmity_container.h
@@ -56,7 +56,7 @@ public:
     void    LogoutReset(uint32 EntityID);		// Sets entry to inactive
     void    AddBaseEnmity(CBattleEntity* PEntity);
     void    UpdateEnmity(CBattleEntity* PEntity, int16 CE, int16 VE, bool withMaster = true, bool tameable = false);
-    void    UpdateEnmityFromDamage(CBattleEntity* PEntity, uint16 Damage);
+    void    UpdateEnmityFromDamage(CBattleEntity* PEntity, int32 Damage);
     void    UpdateEnmityFromCure(CBattleEntity* PEntity, uint8 level, uint16 CureAmount, bool isCureV);
     void    UpdateEnmityFromAttack(CBattleEntity* PEntity,uint16 Damage);
     bool    HasID(uint32 ID); //true if ID is in the container with non-zero enmity level
